pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai>=1.84.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
]
